Rearrange Pins
Rearrange pins to make room for DB-25 header plugs.

The DB25 plugs are too wide to fit with all the pins in place.  I plan to remove 2 pins on the left side of each plug.  The upper plug I will just solder wires to the LEDs, the lower plug I will push the pins down so they are on the on the underside of the board for test points.