Implement full motion detection and add noise supression on step line
Limits are only passed to Smoothie when there is motion in the direction of the limit.

Motion is reported with SSI

Added a count of 128 ad 50MHz (2.5uS) for noise supression.. this was not needed, but seemed like extra insurance.

Changed timing required for reset of SSI to reduce false reads due to reset being too short